data access also affects reproducibility
I suggest some short paragraph (I offloaded a quick idea) just arguing how important it is to have easy access to all the data in an easy manner. It makes it easier to check others results (reproducibility). It could also tie in later in the introduction where we touch on reproducibility (although a different view point).